fix: Improve the smart wallet revival handshake
The wallet factory calls the provision pool wallet reviver reviveWallet
method when receiving a bridge message about a currently-unknown
address, and reviveWallet checks that the address is revivable but not
yet revived to either reject the request or to call back into the
factory's provideSmartWallet method with the correct bank and naming hub
(which the provision pool has but the wallet factory does not).
provideSmartWallet, whether called in this way or more directly, follows
wallet construction with a finisher that informs the provision pool
about the corresponding address, which it removes from the collection of
revivable addresses and then reports back whether or not it was present
(i.e., whether or not it recognizes the invocation as completing a
revival). provideSmartWallet consumes this result to report whether or
not the wallet is actually new.

This increases the count of round trips associated with both new wallet
provisioning (adding one from the wallet factory to the reviver) and
with handling messages about not-yet-known addresses (adding one from
the wallet factory to the reviver for what would previously have been a
more direct failure, and in the case of successful revival [which
previously did not exist at all] one more still from the reviver to the
factory). Since wallet creation is relatively rare and a single internal
round trip is not a particularly attractive vector for attacks whose own
initiation has a message cost, this seems acceptable.