fix: fetch go 1.20.3 but tolerate truncated semver
Gvm cannot fetch the latest patch release of a given minor,
so pick the latest 1.20 release that I've tested.
Go releases can have truncated semver (might there be a "2" release?),
so tolerate that in the regexp and flip the sense of the version check
to fail closed instead of failing open.