feat: add whole-swingset benchmark mode, enabled via --sbench
In this mode, the configured swingset is loaded indirectly, interposing a
swingset-runner-provided benchmark controller as the swingset's bootstrap vat.
The benchmark author is expected to provide a benchmark driver vat in a file
named `vat-benchmark.js` adjacent to the swingset's config file.  The benchmark
driver vat is expected to expose the methods `setup` and `runBenchmarkRound`.
The controller bootstrap invokes the swingset's "real" bootstrap and then tells
the benchmark driver vat to perform setup for the benchmark.  It then serves as
a controller for triggering the execution of the directed number of bootstrap
rounds.

See demo/vaultPerfTest/vat-benchmark.js for an example.