refactor(internal): deprecate makeAggregateError in favor of Aggregat… (#9275)
closes: #XXXX
refs: https://github.com/endojs/endo/pull/2042

## Description

Starting at https://github.com/endojs/endo/pull/2042 endo fully supports
the real `AggregateError` constructor, so there is no longer any need
for the `makeAggregateError` workaround. The @agoric/internal packages
*should* not be imported by anything outside the agoric-sdk repo, so
technically I could remove it in this same PR. As a conservative
don't-break-things, I still export it as deprecated.

Even after upgrading to an endo supporting `AggregateError`, the old
`makeAggregateError` was emulating it using a direct instance of
`Error`, so this is in theory not a pure refactor. But I would be
flabbergasted if any code came to depend on this not being a genuine
`AggregateError`, in which case this change is a pure refactor.

Reviewers, please let me know if you'd prefer that I delete
`makeAggregateError`, as I'd be happy to.

### Security Considerations
none

### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

none
### Upgrade Considerations

none

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>